+++ src/ItemBox/ItemBox.csproj
@@ -1,0 +1,5 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\template.targets" />
+
+</Project>
