+++ src/ItemBox/README.md
@@ -1,23 +1,22 @@
-# itemBox 物品盒子
+# ItemBox 物品盒子
 
-- 作者: 2409 or CAI?
-- 出处: 旧的xsb bot 插件
+- 作者: 2409 & Cai
+- 出处: XSB插件邮件部分
 - 或者叫离线背包更好？但是英文名不知道怎么取诶。
 
 ## 更新日志
 
 ```
-v1.0.0
-初次分离提交。
+v2024.10.3 初次分离提交。
 ```
 
 ## 指令
 
 | 语法           |        权限         |   说明   |
 | -------------- | :-----------------: | :------: |
-| "cbox", "领盒子" | itemBox.claim   | 领取盒子物品|
-| "gbox", "给盒子" |  itemBox.give  | 赠送盒子到指定玩家|
-| "rbox", "清盒子" |  itemBox.remove  | 清空所有盒子|
+| "cbox", "领盒子" | ItemBox.claim   | 领取盒子物品|
+| "gbox", "给盒子" |  ItemBox.give  | 赠送盒子到指定玩家|
+| "rbox", "重置盒子" |  ItemBox.remove  | 清空所有盒子|
 
 ## 配置
 无
